chore: move types to types folder
