+++ packages/connector_sdk/lib/src/types/types.dart
@@ -1,0 +1,2 @@
+export 'connector_error.dart';
+export 'connector_response.dart';
